requirements.txt
python-dotenv
streamlit
opencv-python
numpy
easyocr
moviepy
openai
spacy
chromadb